Units Info Menu and Stage Info UI
The units info menu and the stage info UI have been set up. The visuals need to be worked on,  but mechanically they seem fine. Next I'll set up transferring the necessary information for starting a stage. I might do the tutorial text as well, but that will probably be for later.