Added Unit Button Randomization
Added randomization to the unit buttons so that the values are not always in the same order.

The gameplay info object is getting deleted when going from stage to world, but not for world to stage. I do not know WHY this is happening. I don't see a Destroy() call anywhere. I need to fix this.