Implemented Puzzle Display Fading Text + Different Difficulties
The puzzle system appears to work with different difficulties now. Different difficulties have a different amount of buttons, and they seem to work now.

The fading text mechanic seems to work with the puzzle displays now. The formatting could be improved, but it's all functional now.

The swap puzzle still needs to actually move the symbols around, so that needs to be addressed.