Adjusted Units Info Window (Comparison Bars, TTS) + Char Limit Changes
The comparison bars have been adjusted so that they show the units for the bar fraction.

The TTS for the units info menu has been altered. If the table view is opened, it reads the group description. If the comparison view is opened, it reads the group name. This is because the group description isn't visible in the comparison view.

Some of the character limits have also been changed for the language.json file.