Laser Shot Now Tracks Units Button
The laser shot now tracks the units button. If there are any issues where the value comparisons do not match up, doing a hard check with the units button should correct the problem. The units button tracks if its the button is correct regardless of the value, so checking it should prevent any faulty comparisons that stop a correct answer from registering.